feat: add favorite list and tabbed navigation
Introduces a "Saved" tab to display favorite countries in a dedicated list.

Improves the user experience by providing tabbed navigation
between the country list and the favorite list.

Fixes an issue where the wrong country was selected when tapping on a row.

Standardizes the usage of 'isFavorite' property to reflect the correct state.